+++ src/components/mentoring/dashboard/ContactTabMentorSide.jsx
@@ -14,17 +14,12 @@ import { connectionState } from "../../../constants/mentoring";
 import FlagTag from "../../common/FlagTag";
 import Tag from "../../common/Tag";
 import Button from "../../common/Button";
-import Fallback from "../../common/Fallback";
-import Loader from "../../common/Loader";
-import Error from "../../common/Error";
-import ProfileModal from "./ProfileModal";
+import CreateProfileModal from "./CreateProfileModal";
 
 export default function ContactTabMentorSide() {
   const navigate = useNavigate();
   const queryClient = useQueryClient();
-
   const [isModal, setIsModal] = useState(false);
-
   const [modalUid, setModalUid] = useState(null);
 
   const { data } = useQuery({
@@ -85,7 +80,7 @@ export default function ContactTabMentorSide() {
             onSuccess: () => {
               toast("Successfully accepted.");
               queryClient.invalidateQueries({
-                queryKey: ["post", data.postId],
+                queryKey: ["contacts"],
               });
               setChecks(
                 data.data.data
@@ -114,7 +109,7 @@ export default function ContactTabMentorSide() {
             onSuccess: () => {
               toast("Successfully refused.");
               queryClient.invalidateQueries({
-                queryKey: ["post", data.postId],
+                queryKey: ["contacts"],
               });
               setChecks(
                 data.data.data
@@ -280,13 +275,11 @@ export default function ContactTabMentorSide() {
           Refuse
         </Button>
       </div>
-      <Fallback Loader={Loader} Error={Error} errorMessage="ERROR">
-        <ProfileModal
-          isModal={isModal}
-          setIsModal={setIsModal}
-          uid={modalUid}
-        />
-      </Fallback>
+      <CreateProfileModal
+        isModal={isModal}
+        setIsModal={setIsModal}
+        uid={modalUid}
+      />
     </>
   );
 }
